Replace legacy SplitPanel with v6 SplitView #11013
- Put BrowseLayout into a correct folder according to FSD
- Made preview panel in wizard not collapse when increasing/decreasing details panel size
